requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
langchain>=0.0.346
langchain-openai>=0.0.5
openai>=0.28.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
googletrans>=4.0.0
